chore(test): Use lerna, cleanup commands, split packages
