--- .npmignore
@@ -1,4 +1,0 @@
-!dist/**
-!distForNoAdditional/**
-tests/
-docs/
