+++ package.json
@@ -35,7 +35,7 @@
     "clean": "lerna run clean --stream",
     "prepare": "yarn build",
     "preversion": "yarn test",
-    "test": "yarn --cwd tests test",
+    "test": "lerna run test --stream",
     "lint": "eslint . --ext .ts",
     "lint:fix": "yarn run lint --fix",
     "watch": "lerna run watch --parallel",
@@ -61,11 +61,9 @@
     "node": ">=12.0.0"
   },
   "engineStrict": true,
-  "workspaces": {
-    "packages": [
-      "packages/*",
-      "tests",
-      "tests/esm"
-    ]
-  }
+  "workspaces": [
+    "packages/*",
+    "tests",
+    "tests/esm"
+  ]
 }
