+++ tests/esm/.mocharc.json
@@ -1,8 +1,5 @@
 {
-    "extension": ["ts"],
-	"node-option": [
-		"experimental-specifier-resolution=node",
-		"loader=ts-node/esm"
-	],
-    "spec": ["**/*.spec.ts"]
+  "extension": ["ts"],
+  "node-option": ["experimental-specifier-resolution=node", "loader=ts-node/esm"],
+  "spec": ["**/*.spec.ts"]
 }
