+++ tests/esm/package.json
@@ -5,20 +5,8 @@
   "version": "4.0.0-alpha.1",
   "type": "module",
   "sideEffects": false,
-  "keywords": [
-    "typescript",
-    "openapi",
-    "swagger",
-    "server",
-    "node",
-    "node.js",
-    "codegen",
-    "generation",
-    "express",
-    "hapi.js",
-    "koa"
-  ],
   "scripts": {
+    "pretest": "node --experimental-loader=ts-node/esm prepare.ts",
     "test": "cross-env NODE_ENV=tsoa_test mocha --exit --timeout 5000"
   },
   "author": "Luke Autry <lukeautry@gmail.com> (http://www.lukeautry.com)",
