+++ tests/esm/tsconfig.json
@@ -1,31 +1,22 @@
 {
   "compilerOptions": {
-    "outDir": "dist",
+    "noEmit": true,
     "rootDir": ".",
-    "sourceMap": true,
-    "noImplicitAny": false,
     "baseUrl": ".",
     "module": "esnext",
     "target": "ES2020",
     "lib": ["es2020"],
     "typeRoots": ["node_modules/@types", "../typings"],
-    "downlevelIteration": true,
     "experimentalDecorators": true,
-    "declaration": true,
-    "noUnusedLocals": true,
-    "strictNullChecks": true,
-    "newLine": "LF",
-    "noUnusedParameters": false,
     "emitDecoratorMetadata": true,
-    "noImplicitThis": true,
-    "noImplicitReturns": true,
-    "noFallthroughCasesInSwitch": true,
-    "strictBindCallApply": true,
     "moduleResolution": "node",
     "esModuleInterop": true,
+    "allowSyntheticDefaultImports": true,
+    "strict": true,
     "paths": {
       "@tsoa/cli/*": ["../../packages/cli/src/*"],
-      "@tsoa/runtime/*": ["../../packages/runtime/src/*"]
+      "@tsoa/runtime/*": ["../../packages/runtime/src/*"],
+      "tsoa-tests/*": ["../../tests"]
     }
   },
   "references": [{ "path": "../../packages/runtime" }, { "path": "../../packages/cli" }]
